requirements.txt
fastapi==0.99.0
pydantic
uvicorn
mangum
python-dotenv
selenium
requests
pyclean
pytest
httpx
bs4
requests
fake_useragent